requirements.txt
streamlit
python-dotenv
pillow
google-genai
